requirements.txt
pyrogram
TgCrypto
aiohttp
python-dotenv
dnspython
motor
